refactor : integrate inputs to loginInput
